dev_in_test_app_team/requirements.txt
antiorm==1.2.1
Appium-Python-Client==2.11.1
asgiref==3.4.1
asttokens==2.0.5
attrs==23.1.0
backcall==0.2.0
certifi==2022.6.15.1
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==36.0.1
DateTime==4.7
db==0.1.1
db-sqlite3==0.0.1
decorator==5.1.1
distlib==0.3.6
Django==1.11
django-classy-tags==2.0.0
django-cms==3.9.0
django-formtools==2.3
django-sekizai==2.0.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
docker==6.0.1
exceptiongroup==1.1.3
executing==0.8.3
Faker==15.1.1
filelock==3.8.2
firebase==4.0.1
Flask==2.2.2
h11==0.14.0
idna==3.3
iniconfig==2.0.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
numpy==1.24.3
outcome==1.2.0
packaging==22.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.6.0
pluggy==1.2.0
prompt-toolkit==3.0.30
pure-eval==0.2.2
pycparser==2.21
pygame==2.1.2
Pygments==2.12.0
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.3
pywin32==305
requests==2.28.1
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
source==1.2.0
sqlparse==0.4.2
stack-data==0.3.0
subprocess.run==0.0.8
tomli==2.0.1
traitlets==5.3.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==1.26.12
virtualenv==20.17.1
wcwidth==0.2.5
websocket-client==1.4.2
Werkzeug==2.2.2
wsproto==1.2.0
zope.interface==5.5.1
